Moved code into Homepage component
